requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
environs==8.0.0
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
marshmallow==3.12.1
multidict==5.1.0
Pillow==8.2.0
psycopg2==2.8.6
pybase62==0.4.3
python-dotenv==0.17.1
pytz==2021.1
qrcode==6.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
whitenoise==5.2.0
yarl==1.6.3
